+++ tests/unit/plugins/common/test_types.py
@@ -37,7 +37,8 @@ class PathOrUrlTestCase(test.TestCase):
         self.assertRaises(exceptions.InvalidScenarioArgument,
                           types.PathOrUrl({}, {}).pre_process,
                           "fake_path", {})
-        mock_requests_head.assert_called_once_with("fake_path", verify=False)
+        mock_requests_head.assert_called_once_with(
+            "fake_path", verify=False, allow_redirects=True)
 
     @mock.patch("os.path.isfile")
     @mock.patch("requests.head")